+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/OmBarnetStep.tsx
@@ -21,10 +21,10 @@ import { useSøknadContext } from '../../context/hooks/useSøknadContext';
 import { useStepFormValuesContext } from '../../context/StepFormValuesContext';
 import SøknadStep from '../../SøknadStep';
 import { getSøknadStepConfigForStep } from '../../søknadStepConfig';
-import IkkeHøyereRisikoForFraværAlert from './info/IkkeHøyereRisikoForFraværAlert';
-import IkkeKroniskEllerFunksjonshemningAlert from './info/IkkeKroniskEllerFuksjonshemningAlert';
-import IkkeSammeAdresseAlert from './info/IkkeSammeAdresseAlert';
-import TrengerIkkeSøkeForBarnAlert from './info/TrengerIkkeSøkeForBarnAlert';
+import IkkeHøyereRisikoForFraværAlert from './alert/IkkeHøyereRisikoForFraværAlert';
+import IkkeKroniskEllerFunksjonshemningAlert from './alert/IkkeKroniskEllerFuksjonshemningAlert';
+import IkkeSammeAdresseAlert from './alert/IkkeSammeAdresseAlert';
+import TrengerIkkeSøkeForBarnAlert from './alert/TrengerIkkeSøkeForBarnAlert';
 import { getOmBarnetStepInitialValues, getOmBarnetSøknadsdataFromFormValues } from './omBarnetStepUtils';
 import AnnetBarnFnrSpørsmål from './spørsmål/AnnetBarnFnrSpørsmål';
 import AnnetBarnFødselsdatoSpørsmål from './spørsmål/AnnetBarnFødselsdatoSpørsmål';
