+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/alert/IkkeKroniskEllerFuksjonshemningAlert.tsx
@@ -1,0 +1,12 @@
+import { Alert } from '@navikt/ds-react';
+import { AppText } from '../../../../i18n';
+
+const IkkeKroniskEllerFunksjonshemningAlert = () => {
+    return (
+        <Alert variant="info">
+            <AppText id="steg.omBarnet.alert.ikkeKroniskSykdom" />
+        </Alert>
+    );
+};
+
+export default IkkeKroniskEllerFunksjonshemningAlert;
