+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/alert/TrengerIkkeSøkeForBarnAlert.tsx
@@ -1,5 +1,7 @@
 import { Alert, BodyShort, Heading, Link } from '@navikt/ds-react';
+import { ReactNode } from 'react';
 import { getEnvironmentVariable } from '@navikt/sif-common-core-ds/src/utils/envUtils';
+import { AppText } from '../../../../i18n';
 
 interface Props {
     barnetsFornavn: string;
@@ -9,12 +11,16 @@ const TrengerIkkeSøkeForBarnAlert = ({ barnetsFornavn }: Props) => {
     return (
         <Alert variant="warning">
             <Heading size="small" level="3">
-                Du trenger ikke søke for {barnetsFornavn}
+                <AppText id="steg.omBarnet.trengerIkkeSøke.tittel" values={{ barnetsFornavn }} />
             </Heading>
             <BodyShort>
-                Du har allerede et gyldig vedtak som gjelder til og med det kalenderåret {barnetsFornavn} fyller 18 år.
-                Du trenger derfor ikke å søke på nytt. Du kan finne melding og dokumentasjon om vedtaket på{' '}
-                <Link href={getEnvironmentVariable('MINSIDE_URL')}>Min side</Link>.
+                <AppText
+                    id="steg.omBarnet.trengerIkkeSøke.tekst"
+                    values={{
+                        barnetsFornavn,
+                        Lenke: (child: ReactNode) => <Link href={getEnvironmentVariable('MINSIDE_URL')}>{child}</Link>,
+                    }}
+                />
             </BodyShort>
         </Alert>
     );
