--- apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/info/IkkeSammeAdresseAlert.tsx
@@ -1,9 +1,0 @@
-import { Alert } from '@navikt/ds-react';
-import { useAppIntl } from '../../../../i18n';
-
-const IkkeSammeAdresseAlert = () => {
-    const { text } = useAppIntl();
-    return <Alert variant="info">{text('steg.omBarnet.spm.sammeAdresse.neiAlert')}</Alert>;
-};
-
-export default IkkeSammeAdresseAlert;
