+++ apps/omsorgspengesoknad/src/app/søknad/steps/om-barnet/stegOmBarnetMessages.ts
@@ -65,75 +65,14 @@ const nb = {
         'Beskrivelse på hvordan barnets sykdom eller funksjonshemning gir høyere risiko for fravær fra jobb kan ikke inneholde flere enn 2000 tegn.',
     'steg.omBarnet.validation.høyereRisikoForFraværBeskrivelse.stringIsTooShort':
         'Beskrivelse på hvordan barnets sykdom eller funksjonshemning gir høyere risiko for fravær fra jobb må være på minst 5 tegn.',
+
+    'steg.omBarnet.trengerIkkeSøke.tittel': 'Du trenger ikke søke for {barnetsFornavn}',
+    'steg.omBarnet.trengerIkkeSøke.tekst':
+        'Du har allerede et gyldig vedtak som gjelder til og med det kalenderåret {barnetsFornavn} fyller 18 år. Du trenger derfor ikke å søke på nytt. Du kan finne melding og dokumentasjon om vedtaket på <Lenke>Min side</Lenke>.',
 };
 
 const nn: Record<keyof typeof nb, string> = {
-    'steg.omBarnet.hvilketBarn.spm': 'Kva barn gjeld søknaden?',
-    'steg.omBarnet.hvilketBarn.registrerteBarn': 'Barn registrert på deg:',
-    'steg.omBarnet.hvilketBarn.info': 'Viss du skal søkje for fleire barn, må du sende éin søknad for kvart barn.',
-    'steg.omBarnet.hvilketBarn.født': 'Fødd {dato}',
-    'steg.omBarnet.gjelderAnnetBarn': 'Søknaden gjeld eit anna barn',
-    'steg.omBarnet.annetBarn.tittel': 'Anna barn',
-    'steg.omBarnet.fnr.spm': 'Barnet sitt fødselsnummer/D-nummer',
-    'steg.omBarnet.fødselsdato': 'Barnet sin fødselsdato',
-    'steg.omBarnet.fødselsdato.info': 'Barnet må vere fødd etter {minFødselsdato}',
-    'steg.omBarnet.navn': 'Barnet sitt namn',
-    'steg.omBarnet.relasjon': 'Mi relasjon til barnet',
-    'steg.omBarnet.spm.sammeAdresse': 'Bur du saman med barnet?',
-    'steg.omBarnet.spm.sammeAdresse.ja': 'Ja',
-    'steg.omBarnet.spm.sammeAdresse.jaDeltBosted': 'Ja, barnet har delt fast bostad',
-    'steg.omBarnet.spm.sammeAdresse.nei': 'Nei',
-    'steg.omBarnet.spm.sammeAdresse.hvaBetyrDette': 'Kva er delt fast bostad?',
-    'steg.omBarnet.spm.sammeAdresse.hvaBetyrDette.info':
-        'Viss foreldra til barnet ikkje bur saman, kan dei inngå ei avtale om delt fast bostad etter barnelova §36. Barnet bur då fast med begge sine foreldre.',
-    'steg.omBarnet.spm.sammeAdresse.neiAlert':
-        'Det er berre foreldre som bur saman med barnet som kan få ekstra omsorgsdagar frå NAV. Forelderen som bur saman med barnet kan i nokre tilfelle dele sine omsorgsdagar.',
-    'steg.omBarnet.spm.kroniskEllerFunksjonshemmende': 'Har barnet kronisk/langvarig sjukdom eller funksjonshemming?',
-    'steg.omBarnet.spm.høyereRisikoForFravær':
-        'Har du høgare risiko for fråvær på jobb på grunn av barnet sin sjukdom eller funksjonshemming?',
-    'steg.omBarnet.spm.høyereRisikoForFraværBeskrivelse.tittel':
-        'No treng vi ei skildring frå deg om korleis barnet sin sjukdom eller funksjonshemming gjev høgare risiko for fråvær frå jobb:',
-    'steg.omBarnet.spm.høyereRisikoForFravær.alert':
-        'For å ha rett til ekstra omsorgsdagar på grunn av sjukdom eller funksjonshemming, må det vere ein samanheng mellom barnet sin sjukdom/funksjonshemming og risikoen for høgare fråvær frå jobb.',
-    'steg.omBarnet.alert.ikkeKroniskSykdom':
-        'Denne søknaden om ekstra omsorgsdagar gjeld berre for dei som har barn med kronisk/langvarig sjukdom eller funksjonshemming.',
-    'steg.omBarnet.relasjonTilBarnet.mor': 'Mor',
-    'steg.omBarnet.relasjonTilBarnet.far': 'Far',
-    'steg.omBarnet.relasjonTilBarnet.adoptivforelder': 'Adoptivforelder',
-    'steg.omBarnet.relasjonTilBarnet.fosterforelder': 'Fosterforelder',
-
-    'steg.omBarnet.validation.barnetSøknadenGjelder.noValue':
-        'Du må velje kva barn søknaden gjeld, eller velje at søknaden gjeld eit anna barn.',
-    'steg.omBarnet.validation.barnetsFødselsnummer.fødselsnummerHasNoValue': 'Skriv inn barnet sitt fødselsnummer.',
-    'steg.omBarnet.validation.barnetsFødselsnummer.fødselsnummerIsInvalid':
-        'Du har oppgjeve eit ugyldig fødselsnummer. Kontroller at du har tastet inn rett.',
-    'steg.omBarnet.validation.barnetsFødselsnummer.fødselsnummerIsNot11Chars':
-        'Du har oppgjeve eit ugyldig fødselsnummer. Eit gyldig fødselsnummer består av 11 siffer.',
-    'steg.omBarnet.validation.barnetsFødselsnummer.fødselsnummerIsNotAllowed':
-        'Du har oppgjeve ditt eige fødselsnummer som barnet sitt fødselsnummer. Skriv inn barnet sitt fødselsnummer.',
-    'steg.omBarnet.validation.barnetsFødselsnummer.fødselsnummerAsHnrIsNotAllowed':
-        'Du har oppgjeve eit fødselsnummer som ikkje er tillate.',
-    'steg.omBarnet.validation.barnetsFødselsdato.dateHasNoValue': 'Skriv inn barnet sin fødselsdato.',
-    'steg.omBarnet.validation.barnetsFødselsdato.dateHasInvalidFormat':
-        'Barnet sin fødselsdato er ugyldig. Gyldig format er dd.mm.åååå.',
-    'steg.omBarnet.validation.barnetsFødselsdato.dateIsAfterMax':
-        'Fødselsdato kan ikkje vere etter dagens dato. Skriv inn eller vel dato frå datoveljaren.',
-    'steg.omBarnet.validation.barnetsFødselsdato.barnOver18år': 'Det gis ikkje omsorgsdagar til barn over 18 år.',
-    'steg.omBarnet.validation.barnetsNavn.stringHasNoValue': 'Skriv inn barnet sitt namn.',
-    'steg.omBarnet.validation.barnetsNavn.stringIsTooLong':
-        'Namnet på barnet kan ikkje innehalde fleire enn {maks} teikn.',
-    'steg.omBarnet.validation.søkersRelasjonTilBarnet.noValue': 'Du må velje din relasjon til barnet.',
-    'steg.omBarnet.validation.sammeAdresse.noValue': 'Du må svare ja eller nei på om du bur saman med barnet.',
-    'steg.omBarnet.validation.kroniskEllerFunksjonshemming.yesOrNoIsUnanswered':
-        'Du må svare ja eller nei på om barnet har ein kronisk/langvarig sjukdom eller funksjonshemming.',
-    'steg.omBarnet.validation.høyereRisikoForFravær.yesOrNoIsUnanswered':
-        'Du må svare ja eller nei på om du har høgare risiko for fråvær på jobb på grunn av barnet sin sjukdom eller funksjonshemming.',
-    'steg.omBarnet.validation.høyereRisikoForFraværBeskrivelse.stringHasNoValue':
-        'Skriv inn ei skildring på korleis barnet sin sjukdom eller funksjonshemming gjev høgare risiko for fråvær frå jobb.',
-    'steg.omBarnet.validation.høyereRisikoForFraværBeskrivelse.stringIsTooLong':
-        'Skildringa på korleis barnet sin sjukdom eller funksjonshemming gjev høgare risiko for fråvær frå jobb kan ikkje innehalde fleire enn 2000 teikn.',
-    'steg.omBarnet.validation.høyereRisikoForFraværBeskrivelse.stringIsTooShort':
-        'Skildringa på korleis barnet sin sjukdom eller funksjonshemming gjev høgare risiko for fråvær frå jobb må vere på minst 5 teikn.',
+    ...nb,
 };
 export const omBarnetMessages = {
     nb,
